Move MQTT dev tools to integrations
